pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.39.0",
    "gunicorn>=23.0.0",
    "openai>=1.76.2",
    "psycopg2-binary>=2.9.10",
    "slack-sdk>=3.35.0",
    "twilio>=9.5.2",
    "sqlalchemy>=2.0.40",
]
